+++ react/components/Totalizer/TotalizerLabel.js
@@ -2,7 +2,11 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 const TotalizerLabel = ({ label, mobileScroll }) => (
-  <div className={`c-muted-2 f6 ${mobileScroll ? 'nowrap' : ''}`}>{label}</div>
+  <div
+    data-testid="totalizer-label"
+    className={`c-muted-2 f6 ${mobileScroll ? 'nowrap' : ''}`}>
+    {label}
+  </div>
 )
 
 TotalizerLabel.propTypes = {