+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Dropdown's selected option when value is `undefined`.
 - `ButtonGroup` outline.
 
+### Added
+
+- `Totalizer` testids.
+
 ## [9.112.27] - 2020-03-26
 
 ### Fixed